linux/generic: Test that ikconfig == configfile
For linuxManualConfig, we run make oldconfig. However, the generic
kernels should not have their config change at this point. If it
changed, there's something wrong.

In fact, make oldconfig avoid rewriting the config file if it does not
change, so we can rely on it being an exact match.

Hopefully this means things like the build breaking Rust support
(see #436245) does not happen again.

This depends on IKCONFIG=y, however it is already in our common config,
and thus should only fail if someone uses the generic kernel, *and*
overrides IKCONFIG=n, *and* tries to run this test.