nixos/make-disk-image: populate filesystem with the mkfs.ext4 `-d` option
The `cptofs` tool consistently causes multiple failure among disk image
creation and failure of NixOS tests.

In a desperate attempt to add yet another workaround, if the ext4
filesystem is used for the rootfs, use the `-d` option of the
`mkfs.ext4` command to populate the filesystem without having to use
`cptofs`.