tic-80: use lua5_3_compat instead of lua5_3
This is an alternate build of lua 5.3 that enables (at compile time) backwards
compatibility for some 5.1/5.2 features that are otherwise removed in 5.3.

TIC-80 expects these compatibility shims to be present and some TIC-80 games
require them to function. The normal TIC-80 build enables this in its vendored
copy of lua, but the Nix build turns on `PREFER_SYSTEM_LIBRARIES`, which means
we need to enable the compatibility shims in the Lua package we pass to it.