couchdb: use quickjs instead of spidermonkey
Allows us to drop spidermonkey_91, which reached EOL years ago. We won't
switch to the latest supported spidermonkey version for couchdb(128)
because as of commit its support ends in 3 weeks, so there would be no
use trying to update it again after this. Better to just switch now and
avoid pain in the future.