maintainers: drop gm6k
Not actively maintaining their packages. While there is *some* activity
on GitHub and in the Nixpkgs repo, none of that is related to
maintaining the packages they're listed for.

The contribution guidelines say:

> We would typically look at it if we notice that the author hasn't
reacted to package-related notifications for more than 3 months.

There has been no activity on 5 update PRs to `minetest-mapserver`.
`luci-go` is 10 months outdated, with a lot of activity upstream and
failing to build since March this year.