buildRustCrate: add test for wasm bins with hyphens in name
These are currently broken because buildRustCrate will try to mv from
an underscored name to a hyphenated name, but will only try to move an
unsuffixed name; rustc generates wasm binaries with a .wasm suffix.
This test exercises the broken case and hence presently fails; a
succeeding commit will fix it.