spidermonkey_91: drop
It has been EOL for several years now, and the last dependant(couchdb3)
no longer requires it as of #436932.

I've also removed all conditionals in the common.nix build file that
were for versions older than 115, since we no longer ship any versions
older than that. I would like to remove 115 and 128 soon as well, since
as of commit, they reach end of life in 3 weeks(in fact, 115 is EOL for
all platforms we support afaik)