lib.packagesFromDirectoryRecursive: default to newScope from callPackage
This is based on https://github.com/NixOS/nixpkgs/pull/359984#discussion_r2001178033
and fixes the issue described in https://github.com/NixOS/nixpkgs/pull/437251.

When no scope is provided, we fall back to the scope defined by
callPackage. This way, we can still control whether to use nested
scopes via passing 'newScope' as argument or not, while not failing in
case no 'newScope' is provided.

Signed-off-by: Paul Meyer <katexochen0@gmail.com>