maintainers/README.md: clarify maintainer removal
The wording on how to lose maintainer status implied that this was per
package. While this is true, we also follow the same procedure for
entirely inactive maintainers, dropping them from all packages,
including their maintainer handle.

I'm a friend of making things shorter, this should leave enough room for
interpretation of both.

Since removal of inactive maintainers implies that maintainer-list.nix
is about *active* maintainers, clarify that as well.