plowshare: upgrade spidermonkey dependency; modernize
Spidermonkey 115 goes EOL in 3 weeks from this commit, so we want to
upgrade all dependants before then in preparation. plowshare does not
require a specific version, as it uses the CLI tool, so it's easy to
upgrade. I also modernized the code, using makeBinaryWrapper instead of
makeWrapper, and added a homepage for the package as well.