sqlite: enable limit-on-update/delete
Limit-on-update/delete is useful for, amongst other things, bounding the 
amount of work done in a single SQL statement. It is not enabled by 
upstream by default, but Debian does enable it.

Because the parser generator doesn't receive NIX_CFLAGS_COMPILE, the 
setting has to go through the --enable-update-limit configure flag. 
Maybe in the future, if this derivation moves to e.g. CFLAGS, this 
feature enablement can be made uniform with the others.