sqlite: enable limit-on-update/delete
Limit-on-update/delete is useful for, amongst other things, bounding the 
amount of work done in a single SQL statement. It is not enabled by 
upstream by default, but Debian does enable it.

The 'amalgamation'/autoconf distribution from upstream includes 
pregenerated parser files. Enabling limit-on-update requires changes to 
the parser, so rather than the provided amalgamation we have to go for 
the original source files to get the parser generator to run. 
Conveniently, the pristine source is already used as an input to 
tools.nix, so it might even save a download in some cases!

Because the parser generator doesn't receive NIX_CFLAGS_COMPILE, the 
setting has to go through the --enable-update-limit configure flag. 
Maybe in the future, if this derivation moves to e.g. CFLAGS, this 
feature enablement can be made uniform with the others.