nixos/glitchtip: Make working directory configurable, and writable
Without this change the glitchtip service has a working directory that
is within the nix store and is therefore read-only. This breaks the
ability to upload files to glitchtip, including symbol files, when using
the default local file store.

This removes the home directory from the glitchtip user because without
doing so the cp command will nest a "glitchtip" directory inside the
home directory. The home directory isn't necessary anyway - it was
always empty and not used as a working directory.