buildArmTrustedFirmware: use lib.extendMkDerivation
By switching to `lib.extendMkDerivation`, users can use the "finalAttrs"
style when constructing their ATF builds. In addition, move
buildArmTrustedFirmware to its own callPackage'd file, which allows for
all arguments used to construct `buildArmTrustedFirmware` to be easily
overridable. This was previously difficult and cumbersome to do due to
the buildArmTrustedFirmware enclosing the function arguments in
`pkgs/misc/arm-trusted-firmware/default.nix`, but not being available on
the package-set to override said arguments.