freebsd.mkDerivation: delay stdenv.cc checks
This fixes cross-evaluating the FreeBSD stdenv from Linux.
Previously, it failed due to the pam availableOn check in util-linux,
because trying to check meta.platforms for FreeBSD pam failed, since
it wasn't possible to even compute which attributes were available in
the pam derivation without evaluating stdenv.cc, which was a cyclic
dependency.  By moving these checks inwards, so that only the values
of certain attributes depend on stdenv.cc, not the whole attribute
set, we avoid the infinite recursion.

Fixes: 002e2897e6ee ("util-linux: don't build pam support on static")