facetimehd-calibration: fix build
Resolves #430685

This was the only user of `fetchurl` which was using `-r`, which is incompatible with `--continue-at -` from 4d7a4fbd9f634d0c2f1e660d3981e965b3513466.
Rather than engineer a way to pass range options to `fetchurl`, let's just make a custom builder for this derivation.
It actually makes the whole thing more clear.