haskellPackages.cabal2nix-unstable: set version according to contribution guidelines
The contribution guidelines require an unstable version to be leading
with a real version, or if none is available with `0-`.

This is because certain Nix operations split a package's full name into
name and version component starting with the first section starting with
a digit.

(cherry picked from commit 453d0f8eab3c0fb27968705d7f8a45328b7cc03b)