frigate: centralize the ffmpeg flavor selection in the module
In 0.16.0 frigate added a config option to pass an ffmpeg build for the
ffmpeg and ffprobe executables.

Previously we unconditionally patched in ffmpeg-headless, which has now
been changed to use the `ffmpeg.path` provided build, so that we don't
bloat the closure with multiple ffmpegs or don't require overriding the
package to select another version.

In contrast to the upstream configuration we only accept packages and not
the "default" string, because that would have complicated the logic
needlessly.