tpm2-tss: change default path of fapi config file
The default path is hard-coded at compile time. Previously it pointed to
a location inside the nix store, which made it necessary to use
environment variables to specify the path to a user-specified file.

The nature of the config file is such that it's not possible to specify
one that works for every system this package may be installed in, so
user-configuration is usually necessary.

The default is now /etc/tpm2-tss/fapi-config.json, which
is more compatible with non-nixos config conventions. NixOS
compatibility is now accomplished by linking from the well-known
location to a config file in the nix store.