ghidra: Provide a stable path to yajsw directory
The yajsw tool is used to start Ghidra server, but it's not stored in a
stable path since its version number is attached to the directory name,
which might or might not change any time. So, in order to workaround
that for the upcoming NixOS module, create a symlink in the Ghidra
package.

Signed-off-by: Felix Singer <felixsinger@posteo.net>