haskellPackage.linear-tests: disable flaky (?) test suite
This looks like a badly designed property test suite which fails to
exclude values that can't exhibit correct behavior (NaN), but it could
also theoretically be a correctness issue. Haven't gotten a response from
upstream. Kicking the package a few times usually gets the test suite to
succeed, but I'm kind of sick of that game.