CONTRIBUTING: clarify pushing to others' PRs is allowed
While squash-merges have been discouraged from already, they become
impossible to do when the Merge Queue is enabled. In certain scenarios
it was tolerated to use these, however, thus we introduce an acceptable
replacement instead.

By making explicit that pushing to contributors PRs is allowed by
default, it should be easier for committers to do so.