linuxPackages.evdi: Fix flags
Since #402198 kernel modules are no longer built with wrapped compilers.

Firstly, NIX_CFLAGS_COMPILE is no longer effective, so we change that to
CFLAGS, which is used in the Makefile for evdi.

Secondly, we still need to use the wrapped compiler for userspace stuff,
so separate that out in postBuild.

Fixes #437311