ghidra: Rewrite Ghidra server config file path in scripts
These two scripts are used to administrate the Ghidra server, but the
path to the server configuration file is hard-coded and points to the
example config from the package, which is read-only and may not be
edited by the NixOS module. So, rewrite the path to the /etc directory,
where the NixSO module can put files.

Signed-off-by: Felix Singer <felixsinger@posteo.net>