replaceVars: work around content-addressed output path shenanigans
The placeholder output paths in content-addressed derivations lack the
derivation name, which breaks the general assumption that we could
filter them using said name (e.g. when adjusting patch sets).

This is not really the fault of `replaceVar`, but the combination of it
together with a patch seems common enough that we can just work around
this here.