maintainers: drop gm6k
Although this maintainer has responded to requests for maintainance and
is active in Nixpkgs, the new data collection requirements introduced in
https://github.com/NixOS/nixpkgs/pull/437085 have been a privacy
concern for this maintainer, who has stated (https://github.com/NixOS/nixpkgs/pull/437082#issuecomment-3243483517)
that they do not wish for their data to be added to the maintainer list.

For this reason, there is no other recourse than to remove this
maintainer's information from newer revisions of Nixpkgs, as their
maintainer entry is now non-compliant with Nixpkgs' latest policies.
Once more, this removal is NOT due to the usual inactivity reasons, but
for specific privacy concerns expressed by the maintainer being removed.

Signed-off-by: Fernando Rodrigues <alpha@sigmasquadron.net>