haskell.compiler.ghc{947,948,963,964,965,966,967,981,982,983,984,9101,9102}: backport patches for LLVM support
LLVM 12–17 will be dropped for Nixpkgs 25.11. As discussed recently
on Matrix, this backports upstream changes to allow the use of LLVM
19 for all GHC versions from 9.4.7 onward. I did not bother doing any
work for older GHC versions (other than the 9.0.2 binary, separately),
as they are due for removal anyway.

I looked over GHC commits mentioning LLVM since the release of 9.4.7,
and read the discussions and issues around the relevant bumps, and
attempted to be quite thorough, but I obviously cannot guarantee that
this is wholly comprehensive. It seems like upstream generally bumps
the upper bound on the basis of “it builds successfully for me”,
with specific adaptations for new versions being fairly uncommon and
only coming for obvious build blockers or reactively in response to
bug reports. I have backported both kinds of changes here.

For some commits, trivial conflict resolutions and adaptations
were required. It would be possible to pass the affected files to
`fetchpatch` as `excludes` and keep smaller fix‐up patches in tree in
some cases, but I opted to keep it simple and vendor complete backport
patches instead. I did not attempt to backport every single change to
the LLVM backend, only those that seemed directly relevant to support
for newer versions; if you’d get the same issue with the older LLVM,
that’s just a GHC bug.

These changes should actually make it easier to cross‐compile for
new architectures, as more recent LLVMs will have better support for
newer platforms, and it will be easier to backport GHC changes to
enable new platforms with less drift in the backend.

These patches do result in two breaking changes. Firstly, the minimum
LLVM version is bumped to 13 across the board. This is irrelevant for
Nixpkgs as we pin a specific LLVM version anyway, and versions below
LLVM 18 will be removed imminently. Secondly, support for the hidden
`-fno-llvm-tbaa` flag is dropped. This can be replaced with custom
`-optlo` flags to control the passes more directly, but the main
use of this undocumented flag appears to have been to [work around]
the lack of support for newer LLVM versions, anyway.

[work around]: <https://gitlab.haskell.org/ghc/ghc/-/issues/22220>

I successfully built the following on `aarch64-linux`:

* `haskell.compiler.ghc928`
* `haskell.compiler.ghc947`
* `haskell.compiler.ghc948`
* `haskell.compiler.ghc963`
* `haskell.compiler.ghc981`
* `haskell.compiler.ghc9101`
* `haskell.compiler.ghc9102`
* `pkgsCross.riscv64.haskell.compiler.ghc948`
* `pkgsCross.armv7l-hf-multiplatform.buildPackages.haskell.compiler.ghc964`
* `pkgsCross.armv7l-hf-multiplatform.buildPackages.haskell.compiler.ghc965`
* `pkgsCross.armv7l-hf-multiplatform.buildPackages.haskell.compiler.ghc966`
* `pkgsCross.armv7l-hf-multiplatform.buildPackages.haskell.compiler.ghc967`
* `pkgsCross.armv7l-hf-multiplatform.buildPackages.haskell.compiler.ghc982`
* `pkgsCross.armv7l-hf-multiplatform.buildPackages.haskell.compiler.ghc983`
* `pkgsCross.armv7l-hf-multiplatform.buildPackages.haskell.compiler.ghc984`

The 9.6 and 9.8 minor versions that I built natively were already
broken for ARMv7 cross due to [GHC bug #24349]. GHC 9.10 also failed;
I don’t know what the cause is (`armv7l-unknown-linux-gnueabihf-ghc:
could not execute: clang` is not very helpful), but it also occurs
on `master`. Perhaps [the manual] is out of date and Clang is now
required to use the LLVM backend even on non‐Darwin hosts? I tried
them with `pkgsCross.s390x`, but the same error occurred on 9.10,
and GCC was unhappy about unknown instructions in assembly code for
9.6 and 9.8. Anyway, I consider these issues to be out of scope for
this change, but perhaps unconditionally using Clang with the LLVM
backend might help.

[GHC bug #24349]: <https://gitlab.haskell.org/ghc/ghc/-/issues/24349>
[the manual]: <https://downloads.haskell.org/ghc/latest/docs/users_guide/codegens.html#llvm-code-generator-fllvm>

I have built a test program with the ARMv7 cross‐compilers and
confirmed that they run under `qemu-arm(1)`. I also confirmed
that the cross‐compiled RISC‐V GHC successfully runs under
`qemu-riscv64(1)`. It will only try to build programs via the C
backend, though, as that is the only option for unregisterised™
targets, so it’s not clear to me how useful LLVM support in 9.4.8
really is for bootstrapping new platforms; I guess even RISC‐V
would require more backporting work to produce a cross‐compiled
GHC that will use LLVM to compile its own input. I didn’t bother
setting up all the binfmt machinery to get it through compiling and
running a test program, but it at least makes the attempt.