nixosTests.simple: drop minimal profile
This should be a test that will change whenever most likely *all* NixOS
tests change. As such it should use the most "default" configuration
possible. This decreases chances of a change that will affect almost
other NixOS tests to go unnoticed by this test.

With the minimal profile used previously, there could be changes made to
the default behavior of NixOS modules which were explicitly disabled in
this profile. These would not change the simple test, but would change
all other tests.

Removing as many moving parts as possible from this test should make it
the most default it can be. It now contains only pieces that every other
NixOS test has, too: A name, a machine and a testScript. But nothing
fancy.