floorp-bin: init at 12.1.4
Fixes #438494.
Fixes #431292.

It's a -bin variant of floorp, which previously we built from source.
Upstream became pretty much unfeasible to build from source starting
with the 12.x release line, so let's replace it with a `-bin` variant.

Long-term this makes it also a lot easier on maintenance and Hydra,
both resource and time-wise.

Signed-off-by: Christoph Heiss <christoph@c8h4.io>