buildArmTrustedFirmware: add override functionality for backward-compat
This is just to allow for backward-compatibility of existing
buildArmTrustedFirmware derivations. The downside of the original
implementation was that you could not update `prevAttrs`, since
`overrideAttrs` was not being used. This shim essentially provides the
same functionality since it calls overrideAttrs without referencing
`prevAttrs`, and thus it blindly overrides the derivation attrs.