haskell.packages.native-bignum: exclude recurseForDerivations
This was accidentally introduced when it was added to
`haskell.compiler`. This code is a bit odd, because it iterates through
the compiler attributes and then tries to access the same name in
packages, which fails in this case.

Seems like this was just a copy-paste error, this should iterate through
packages instead.

(cherry picked from commit 07cb32debe7052211fa60aa6bb2f991f57b7c626)