haskell.compiler.ghc865Binary: drop
This was used for bootstrapping GHC on powerpc64le, but it's unclear
whether that actually still works. It makes no sense to pretend to
support it, when it's most likely broken, but we can't even verify.

Also the bootstrap chain will be broken once we drop GHC 8.10.7, so we
might as well drop the whole thing.