nixos/tests: RFC: change `Machine.systemctl` to use `Machine.succeed`
Previously we used `Machine.execute`, which doesn't actually check that
the command succeeded (it just returns the (status code, output).

I grepped around and didn't find anywhere in nixpkgs where we do
anything with a failing `Machine.systemctl` call. We did have a number
of places that explicitly check the result and fail the test, which now
no longer needs to happen.