jupyterLib: init
Initial effort to provide a better interface to jupyter in nixpkgs.
There has been effort with jupyterWith then jupyenv but having some more
foundations in nixpkgs could cover 80% usecases.

See https://github.com/NixOS/nixpkgs/pull/278315 for a first effort.

The starting point for me was when nixpkgs failed evaluating because I
had not provided logo32 and logo64.
The evaluation failure was somewhere in a nix code which is kinda scary. 
Providing an interface/abstraction could help with documentation and
renaming of arguments to be closer to jupyter (I am thinking
interruptMode vs interrupt_mode ?).