nota: 1.0 -> 1.0-unstable-2023-03-01
Homepage: https://github.com/pouyakary/Nota

The url in the fetchurl function no longer exists, the website domain is
for sale. The same holds for the homepage listed. Therefore, the urls
have been updated to the current versions.

The description has been updated to align with the README for pkgs
contributions, namely, that the description should "Avoid subjective
language."