maintainers/scripts/haskell/exclude.nu: init
This script populates `excluded.yaml`, which allows hackage2nix to skip
generating expressions for many old packages.

The script uses a very simple logic right now. Any packages that is both
broken and has not been updated for 5 years on hackage, is filtered out
entirely. An exception is made for packages which are listed in the
current Stackage snapshot as well - these are kept, otherwise
hackage2nix will complain, because it can't deal with the constraints on
these packages.

This does not look at reverse dependencies, so will very likely in
changes to hackage-packages.nix, where some packages that are still left
in the set are passed `null` for their removed dependencies. I'd argue
that this is not a problem per se: These packages were transitively
broken before and would now fail to build because of *missing*
dependencies if anyone tried to resurrect them. As part of fixing these
packages, the contributor would have to fix that dependency anyway,
they'd now need to remove it from the exclusion list to do so.

Not considering reverse dependencies makes the logic considerably
simpler.

Another thing this does not consider is when an already excluded package
is resurrected on hackage, i.e. gets a new upload. This is left out for
now on purpose - should we hit these cases more often, we can still
write a script to automate removal from this list as well.