maintainers: remove federicoschonborn
*This is a very disorganized rant, there's a lot of things going on in my mind when writing this. Proceed with caution.*

I was going to make a short message to explain why I'm removing my Microsoft GitHub account and removing myself from Nixpkgs, which is the only project I had been contributing left here (maybe if you keep integrating more of Microsoft GitHub's crap on this repository it will help migrate somewhere else one day!)

I'm tired of Microsoft GitHub becoming more and more open to associating with fascists, I've been guilty myself for keeping using it for so long (the company's involvement with ICE was known for a long time, then more recently the inclusion of <the company named after Elon Musk's failed PayPal rebranding>'s "xAI" slop generator also known as "Mecha Hitler").

Not to count all the Microsoft crap.

This site is going the way of the old SourceForge.

The thing is, I *loved* Nix, NixOS and Nixpkgs a lot, and I wanted to keep contributing.

This is the third time I'm leaving and this time it's not because of Determinate Systems and other associated companies. Woo-hoo!

I'm sure all those fancy new features that keep being added to the fork are going to trickle down to NixOS/nix *one day*!

But then I said, "I haven't looked at the Discourse for a while, maybe I should take a look at it, *what could possibly go wrong?*"

It turns out, [we did it again](https://discourse.nixos.org/t/sc-member-tomberek-works-for-anduril/68971), Anduril is back to being involved in a leadership position within the NixOS community! Congratulations, Mr. Luckey, you found a loophole!

The Steering Committee proves once again that they are a failed experiment, unfit to lead responsibly. After all the promises that were made we keep failing for the same old tricks.

I'm tired of the long discussions that go nowhere.

I'm tired of seeing people compromising with fascists.

I'm tired of seeing *myself* compromising with fascists.

Fuck this, fuck Microsoft, fuck GitHub, fuck Anduril, fuck Determinate Systems, fuck all those who are helping them inside and outside this community because they're either fascists themselves or because they're incredibly naïve.

Hell,

Fuck *myself* for standing with this shit for so long.

Fuck *myself* for coming back time and time again, knowing what I was doing.

Fuck *myself* for trying to pretend all this shit was going to stop at some point.

Fuck *myself* for trusting people that never really wanted to change any of this.

Fuck *myself* for helping war-profiteering companies thrive, directly or indirectly.

We don't need to meet in the middle with people that want me and a lot of other contributors dead in a ditch.

I'll keep an eye on the Lix and Auxolotl projects and Guix, since their communities have a little more self-respect.

See you on the other side!

Signed-off-by: Federico Damián Schonborn <federicoschonborn@disroot.org>