frigate: add converted OpenVINO SSDLite MobileNetV2 model
The Frigate package now also includes the converted OpenVINO SSDLite
MobileNetV2 detection model, same as the official Frigate Docker image
does. During build, it downloads the SSDLite MobileNetV2 TensorFlow
model and converts it to OpenVINO format using the modern (non-
deprecated) openvino.convert_model() API.

This implementation does not use the suggestion from the previous TODO
comment and does not reuse the script that Frigate uses to convert the
model for the official Docker image. This way it avoids using the
deprecated openvino-dev Python package, which is also not available in
nixpkgs.

The model can now be referenced in the Frigate settings like this:

```
detectors = {
  ov_0 = {
    type = "openvino";
    device = "GPU";
  };
  # ...
};

model = {
  width = 300;
  height = 300;
  input_tensor = "nhwc";
  input_pixel_format = "bgr";
  path = "${pkgs.frigate}/share/frigate/ssdlite_mobilenet_v2_fp16.xml";
  labelmap_path = "${pkgs.frigate}/share/frigate/coco_91cl_bkgr.txt";
};
```

Closes #401399