haskellPackages: mark previously transitively broken as broken
These packages were previously marked as transitively broken, before the
exclusion of old packages temporarily marked them unbroken. I only
tested the first 10 or so, but they all failed to build. Chances are
high, that *all* of them will fail to build, because some of their
dependencies were removed.

In the odd case that one of these would succeed, this would show up in
the next run of unbreak.nu.