tests.haskell.ghcWithPackages: verify -fllvm works with useLLVM
Note that this test also succeeds in pkgsCross though using a stdenv
with cc is crucial, since its setup hooks are required for -lgmp to work.

Maybe this is something we should look at for GHC again, maybe we can
add extra entries to the package db, so it passes the necessary -L flags
by itself?