gcc: always apply Darwin patches for Darwin targets
The upstream issue was fixed, the branch has changes relevant to
`x86_64-darwin` too, and you don’t really want to target Darwin
without these patches.

We reorder the deployment target patch after the patches, as they touch
the same file, and apply it on AArch64 for all versions (but for Darwin
targets rather than hosts, which should be the relevant case anyway).

Co-Authored-By: sempiternal-aurora <78790545+sempiternal-aurora@users.noreply.github.com>