blender: use release config
This ensures that we at least default to the config that upstream
Blender uses. Remove any config that we have that is redundant with [the
current upstream][1]. Need to explicitly turn off
`WITH_CYCLES_CUDA_BINARIES` in non-CUDA (as release build will turn it
on) and turn off `WITH_CYCLES_DEVICE_ONEAPI` (as it was previously
implicitly off and we do not currently support it).

[1]: https://projects.blender.org/blender/blender/src/tag/v4.5.2/build_files/cmake/config/blender_release.cmake