hiawatha: drop
Hiawatha is no longer actively maintained by the upstream developer and
there was no maintainer for this package since the original contributor
was removed in 2019.

With https://kb.cert.org/vuls/id/461364 there are three new CVEs that
will eventually be fixed, but the interest in nixpkgs is likely low
since we don't even ship a module.