nixos/filesystems/ext: Enable by default with systemd initrd
NixOS has traditionally enabled the `ext` family of file systems by
default. Originally, when switching to systemd initrd, we wanted to
transition to making this explicit so that initrds could be made
without `ext`. The problem is that anyone with `fsType = "auto";` for
an `ext` file system in initrd will fail to boot, which is not really
an acceptable regression as we switch to systemd initrd by default.

At the time, we didn't have the ability to set
`boot.initrd.supportedFilesystems.ext4 = false;` like we do now. With
the new attr-style `supportedFilesystems`, anyone who wants to get
`ext` out of their initrd can explicitly disable it.

In hindsight, this is probably what #225352 was really about.