llvmPackages.libclc: expose regardless of version
It doesn’t make sense to drop the attribute just because we
don’t yet have a `spirv-llvm-translator` that can work with it
yet, especially as they’ve been adding support for the native LLVM
SPIR‐V backend. It will now be marked as broken transitively if
`spirv-llvm-translator` is unavailable.