lixPackageSets.lix_2_92: drop
After discussing with the nixpkgs developers and Lix developers, we came
to the conclusion that we could also drop directly 2.92 as this is not a
much used version and 2.93 has no known regression over 2.92 (only
advantages).

To reduce the maintenance churn on the Lix team, we drop this version
for 25.11 and will maintain only for the lifetime of 25.05.

Change-Id: I4ee310df31d36b92e394ff2dfebd589d10b12228
Signed-off-by: Raito Bezarius <masterancpp@gmail.com>