workflows/{merge_group,pr}: fail status check explicitly
When the merge queue fails, the workflow currently does not post a
negative result - and GitHub Actions waits for the status check to time
out, which takes 60 minutes.

This, of course, is a waste of time and resources. By explicitly failing
the status check, we boot the PR out of the merge queue immediately.

(cherry picked from commit 43a8f20d581f1395f2947810b5bfcb270754b9c1)