nixos/default.nix: Allow specialArgs to be passed
To allow plain nixos configrations the same power as flakes, we must
expose specialArgs, otherwise a classic pattern importing all of ones
"inputs" (i.e. npins sources), and then exposing them to the module
system becomes impossible due to the fact that imports must be
statically resolvable so these sources cannot be passed in the existing
entry point using _module.args if the sources themselves have modules to
be imported.